Fix API version bug
The login URL should not include the leading v in the api version, just the numeric part. 

(See docs at: https://developer.salesforce.com/docs/atlas.en-us.api.meta/api/sforce_api_calls_login.htm)

Including the leading v (as is the default behavior with the value "v29.0" results in salesforce returning a server_url with a version 7.0, which  is being retired early next year. I've opted to leave the default value as is,  to avoid changing the library's interface, and just to change the implementation to generate the correct URL.